+++ R/art_studio.R
@@ -1,5 +1,9 @@
-# igraph ------------------------------------------------------------------
+###########################################################################
+# ART STUDIO
+# graph visualization
+###########################################################################
 
+# igraph ------------------------------------------------------------------
 
 #' make igraph
 #'
@@ -21,20 +25,24 @@ mapper_object_to_igraph <- function(mapperobject) {
   vertices = mapperobject[[1]]
   edges = mapperobject[[2]]
 
-  if (length(edges$source) <= 1) {
-    if (edges$source == "") {
+  # deal with edgeless graphs
+  if (length(edges$source) <= 1 & any(edges$source == "")) {
       mappergraph = graph_from_data_frame(d = data.frame(id = rownames(vertices), id = rownames(vertices)),
                                           vertices = vertices)
       mappergraph = delete_edges(mappergraph, E(mappergraph))
-    }
   } else {
     mappergraph = graph_from_data_frame(d = edges,
                                         directed = FALSE,
                                         vertices = vertices)
   }
 
+  # don't label nodes
   mappergraph = set_vertex_attr(mappergraph, "label", value = NA)
+
+  # set node size to square root of cluster size
   mappergraph = set_vertex_attr(mappergraph, "size", value = sqrt(vertices$cluster_size))
+
+  # color nodes if binning
   if ("bin" %in% colnames(vertices)) {
     num_bins = max(vertices$bin)
     colfunc = colorRampPalette(c("blue", "gold", "red"))