+++ NEWS.md
@@ -1,3 +1,8 @@
+# mappeR 1.2.0
+
+* Add other hierarchical clustering methods (those available from `fastcluster`)
+* Fix issues with 100 percent overlap situtations caused by `mapply` simplifications
+
 # mappeR 1.1.0
 
 * Added safety checks for mapper input functions (no `NA` inputs, etc)